refactor card preferences out of preferences
now preferences view can be a container
